Refactor unlock ui classes
